+++ src/main/java/com/smart/tailor/repository/MaterialRepository.java
@@ -31,10 +31,10 @@ public interface MaterialRepository extends JpaRepository<Material, UUID> {
     List<Material>findAllMaterialByExpertTailoringIDAndCategoryID(UUID expertTailoringID, UUID categoryID);
 
     @Query(value = "SELECT MIN(bm.brand_price) FROM brand_material bm join material m on m.material_id = bm.material_id where m.material_id = ?1", nativeQuery = true)
-    Double getMinPriceByMaterialID(UUID materialID);
+    Integer getMinPriceByMaterialID(UUID materialID);
 
     @Query(value = "SELECT MAX(bm.brand_price) FROM brand_material bm join material m on m.material_id = bm.material_id where m.material_id = ?1", nativeQuery = true)
-    Double getMaxPriceByMaterialID(UUID materialID);
+    Integer getMaxPriceByMaterialID(UUID materialID);
 
     boolean existsByMaterialNameIgnoreCaseAndCategory_CategoryNameIgnoreCaseAndHsCodeAndUnitIgnoreCaseAndBasePrice(
             String materialName,
